Moved opal accessor
